Moved common function crate to src folder
